requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2

# Scheduling
APScheduler==3.10.4